Moved object definitions to grpc
